+++ backend/src/backend/gen_insn_selection.hpp
@@ -140,9 +140,9 @@ namespace gbe
     /*! Number of sources */
     uint8_t srcNum:4;
     /*! To store various indices */
-    uint16_t index;
+    uint32_t index;
     /*! For BRC/IF to store the UIP */
-    uint16_t index1;
+    uint32_t index1;
     /*! instruction ID used for vector allocation. */
     uint32_t ID;
     /*! Variable sized. Destinations and sources go here */
